+++ myBlog/enroll-server/src/main/java/com/gcnbl/service/UserService.java
@@ -3,7 +3,7 @@ package com.gcnbl.service;
 public interface UserService {
     int validateName(String username);
 
-    void addBlogUser(String username, String password, int telphone);
+    void addBlogUser(String username, String password, Long telphone);
 
     Long login(String username, String password);
